Handle resuming and pulling new tools (#4)
This update introduces improvements to the dev container setup by
restructuring the logic for tool installation and execution:

1. **Tool Installation and Updates:**
- Moved the logic for installing tools and executables and the CLI
extension from the `onCreateCommand` to a dedicated `refreshTools`
script.
- The `refreshTools` script is now invoked during the `postStartCommand`
(instead of `onCreateCommand`, ensuring that tools are updated and
initialized at both create and resume time.
- `refreshTools` uses marker files to ensure that we don't keep
repulling tools if the version hasn't changed

2. **Tool Invocation:**
- Tool starting is moved from the devcontainer.json into a new
`postStartCommand` script
- This ensures that the tool refresh always finishes before the tools
are started