Fix calendar.tsx typescript errors
We were passing `IconLeft` and `IconRight` to the `react-day-picker` component, when those aren't actually options in the components prop. Per https://daypicker.dev/api/type-aliases/CustomComponents, it looks like what we should've used is PreviousMonthButton and NextMonthButton. Confirmed in a prod codespace editing calendar.tsx directly that this does render the correct icons