Test creating Commit (#117)
We have encountered an issue where the command git rev-parse HEAD fails
for users with GPG signing enabled. This appears to be caused by the
initial commit failing with a 403 error when attempting to sign the
commit, as GPG signing requires push access to the repository. To
address this and ensure an initial commit is always present, we will
implement a check to verify whether a commit has been created; if not,
we will proceed to create one.